[backend] Move all the CLI code to a single module
This change finishes the migration to a god library where
the common code is now part of the CLI module.
