+++ haskell/monocle.cabal
@@ -92,9 +92,7 @@ common codegen
 
 common cli-options
   hs-source-dirs:      app
-  build-depends:       optparse-generic           < 1.5
-                     , monocle
-                     , relude                     > 1.0 && < 1.1
+  build-depends:       base, monocle
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N -with-rtsopts=-T
 
 library
@@ -117,6 +115,7 @@ library
                      , dhall-yaml                 ^>= 1.2
                      , directory
                      , either                     ^>= 5
+                     , envparse                   < 0.4.2
                      , exceptions                 < 0.11
                      , fakedata                   >= 1.0
                      , fast-logger
@@ -135,6 +134,7 @@ library
                      , mtl
                      , network                    < 4
                      , network-uri                < 2.6.5
+                     , optparse-generic           < 1.5
                      , parser-combinators         < 1.3
                      , prometheus-client          ^>= 1.0
                      , prometheus-metrics-ghc     ^>= 1.0
@@ -168,6 +168,7 @@ library
   exposed-modules:     Monocle.Prelude
                      , Monocle.Env
                      , Monocle.Class
+                     , CLI
 
                      -- monocle api
                      , Monocle.Api
@@ -232,12 +233,7 @@ library
 
 executable monocle-api
   import:              common-options, cli-options
-  hs-source-dirs:      app
   main-is:             Api.hs
-  build-depends:       base
-                     , monocle
-                     , relude
-                     , envparse                   < 0.4.2
 
 executable macroscope
   import:              common-options, cli-options
@@ -246,7 +242,6 @@ executable macroscope
 executable lentille
   import:              common-options, cli-options
   main-is:             LentilleStandalone.hs
-  build-depends:       streaming
 
 benchmark json-decode
   import:              common-options
